backend-core/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.5
xlsxwriter==3.2.0
psutil==5.9.8
supabase==2.6.0
python-jose==3.3.0
